+++ functions.php
@@ -298,6 +298,10 @@ add_action( 'admin_init', 'redirect_admin' );
  * Add your own scripts and styles below.
  */
 function ignition_scripts() {
+
+	//load regular versions if script debug is set to true in wp-config file.
+	$suffix = ( defined( 'SCRIPT_DEBUG' ) && SCRIPT_DEBUG ) ? '' : '.min';
+
 	// Add google fonts
 	wp_enqueue_style( 'ignition-fonts', ign_google_fonts_url(), array(), wp_get_theme()->get( 'Version' ) );
 
@@ -305,7 +309,7 @@ function ignition_scripts() {
 	wp_enqueue_style( 'ignition-style', get_stylesheet_uri(), '', wp_get_theme()->get( 'Version' ) );
 
 	//Sass compiles styles. Will get child's theme version if found instead. Child theme should import with sass.
-	wp_enqueue_style( 'ignition-sass-styles', get_theme_file_uri( '/dist/main.min.css' ), '', wp_get_theme()->get( 'Version' ) );
+	wp_enqueue_style( 'ignition-sass-styles', get_theme_file_uri( '/dist/main' . $suffix . '.css' ), '', wp_get_theme()->get( 'Version' ) );
 
 	//ie11 js polyfills
 	wp_enqueue_script( 'polyfill', 'https://polyfill.io/v3/polyfill.min.js?flags=gated&features=AbortController%2Cdefault%2CNodeList.prototype.forEach' );
@@ -318,7 +322,7 @@ function ignition_scripts() {
 
 
 	//any javascript file in assets/js that ends with custom.js will be lumped into this file.
-	wp_enqueue_script( 'ignition-custom-js', get_template_directory_uri() . '/dist/custom.min.js', array( 'jquery' ),
+	wp_enqueue_script( 'ignition-custom-js', get_template_directory_uri() . '/dist/custom' . $suffix . '.js', array( 'jquery' ),
 		wp_get_theme()->get( 'Version' ), true );
 
 	//AJAX ready for .custom.js files
@@ -351,10 +355,14 @@ add_action( 'wp_enqueue_scripts', 'ignition_scripts' );
  * Add Stylesheet for Gutenberg
  */
 function ign_gutenberg_styles() {
-	// Load the theme styles within Gutenberg.
-	wp_enqueue_style( 'ign-gutenberg-style', get_theme_file_uri( '/dist/gutenberg-editor-style.min.css' ), false, wp_get_theme()->get( 'Version' ), 'all' );
+	//load regular versions if script debug is set to true in wp-config file.
+	$suffix = ( defined( 'SCRIPT_DEBUG' ) && SCRIPT_DEBUG ) ? '' : '.min';
 
-	wp_enqueue_script( 'ignition-custom-js', get_template_directory_uri() . '/dist/custom.min.js', array( 'jquery' ),
+
+	// Load the theme styles within Gutenberg.
+	wp_enqueue_style( 'ign-gutenberg-style', get_theme_file_uri( '/dist/gutenberg-editor-style' . $suffix . '.css' ), false, wp_get_theme()->get( 'Version' ), 'all' );
+
+	wp_enqueue_script( 'ignition-custom-js', get_template_directory_uri() . '/dist/custom' . $suffix . '.js', array( 'jquery' ),
 		wp_get_theme()->get( 'Version' ), true );
 }
 
@@ -365,7 +373,9 @@ add_action( 'enqueue_block_editor_assets', 'ign_gutenberg_styles' );
  * Add login stylehseet
  */
 function my_login_styles() {
-	wp_enqueue_style( 'custom-login', get_stylesheet_directory_uri() . '/dist/login-style.min.css' );
+	//load regular versions if script debug is set to true in wp-config file.
+	$suffix = ( defined( 'SCRIPT_DEBUG' ) && SCRIPT_DEBUG ) ? '' : '.min';
+	wp_enqueue_style( 'custom-login', get_stylesheet_directory_uri() . '/dist/login-style' . $suffix . '.css' );
 }
 
 add_action( 'login_enqueue_scripts', 'my_login_styles' );
@@ -373,10 +383,11 @@ add_action( 'login_enqueue_scripts', 'my_login_styles' );
 
 /**
  * Add admin stylesheet
- * experimental
  */
 function footer_styles() {
-	wp_enqueue_style( 'ignition-admin-styles', get_theme_file_uri( '/dist/admin.min.css' ), '', wp_get_theme()->get( 'Version' ) );
+	//load regular versions if script debug is set to true in wp-config file.
+	$suffix = ( defined( 'SCRIPT_DEBUG' ) && SCRIPT_DEBUG ) ? '' : '.min';
+	wp_enqueue_style( 'ignition-admin-styles', get_theme_file_uri( '/dist/admin' . $suffix . '.css' ), '', wp_get_theme()->get( 'Version' ) );
 }
 
 add_action( 'admin_footer', 'footer_styles', 99 );